+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: bento </var/spool/mail/bento>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/26 12:04:30 by bento             #+#    #+#             */
-/*   Updated: 2024/05/22 14:15:09 by bthomas          ###   ########.fr       */
+/*   Updated: 2024/05/23 14:58:34 by bthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int		ft_isascii(int i);
 int		ft_isdigit(int i);
 int		ft_isprint(int i);
 void	*ft_memset(void *s, int c, size_t n);
-int		ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 int		ft_tolower(int i);
 int		ft_toupper(int i);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
